User commit

+++ src/main/java/com/example/demo/repository/StudentRepository.java
@@ -2,6 +2,7 @@ package com. example.demo.repository;
 import org. springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.Student;
 @Repository 
-public interface StudentRepository extends JpaRepository<Student,Long>{
+public interface StudentRepository extends JpaRepository<Student,Integer>{
+    
 
 }
